+++ bin/commands/cpp/test-main.js
@@ -1,8 +1,8 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 const title_1 = require("./title");
-function getTemplateTestMain() {
-    const test_main = `${title_1.getTemplateTitle("test.main.cpp")}\n
+function getTemplateTestMain(options) {
+    const test_main = `${title_1.getTemplateTitle("test.main.cpp", options.config)}\n
 // Add header to your project here.
 
 // Required headers for Micro Test.