+++ src/commands/cpp/source.ts
@@ -1,9 +1,9 @@
 "use strict";
 import { getTemplateTitle } from "./title";
 
-export function getTemplateSource( filename: string ): string {
+export function getTemplateSource( filename: string, options: any ): string {
 
   const header_file = filename.replace( /\.cpp$/i, ".hpp" );
-  const source = `${ getTemplateTitle( filename ) }\n\n#include "${ header_file }"\n`;
+  const source = `${ getTemplateTitle( filename, options ) }\n\n#include "${ header_file }"\n`;
   return source;
 }