+++ src/commands/create-cpp-project.ts
@@ -18,13 +18,13 @@ This C++ Project was generated using Dev Mentor Tools (${ VERSION }).
 Initial Commit.
 `;
 
-function safeCreateHeaderFile( filename: string ): void {
+function safeCreateHeaderFile( filename: string, options: any ): void {
   if ( fs.existsSync( `./src/${ filename }` ) ) {
     console.log( "Header file already exist, skipping!" );
     return;
   }
 
-  fs.writeFile( `./src/${ filename }`, getTemplateHeader( filename ), err => {
+  fs.writeFile( `./src/${ filename }`, getTemplateHeader( filename, options.config ), err => {
     if ( err ) {
       console.log( err.message );
       return;
@@ -33,13 +33,13 @@ function safeCreateHeaderFile( filename: string ): void {
   console.log( `Created Header file: ${ filename }` );
 }
 
-function safeCreateSourceFile( filename: string ): void {
+function safeCreateSourceFile( filename: string, options: any ): void {
   if ( fs.existsSync( `./src/${ filename }` ) ) {
     console.log( "Source file already exist, skipping!" );
     return;
   }
 
-  fs.writeFile( `./src/${ filename }`, getTemplateSource( filename ), err => {
+  fs.writeFile( `./src/${ filename }`, getTemplateSource( filename, options.config ), err => {
     if ( err ) {
       console.log( err.message );
       return;
@@ -71,37 +71,41 @@ export function createCppProject( cmd: any, options: any ): void {
     cmd.cpp.forEach( ( file: string ) => {
       const header_file = file.replace( /\.cpp$/i, ".hpp" );
       header_files.push( header_file );
-      safeCreateHeaderFile( header_file );
-      safeCreateSourceFile( file );
+      safeCreateHeaderFile( header_file, options );
+      safeCreateSourceFile( file, options );
     } );
   }
 
-  fs.writeFile( "./src/main.cpp", getTemplateMain( header_files ), err => {
+  fs.writeFile( "./src/main.cpp", getTemplateMain( header_files, options ), err => {
     if ( err ) {
       console.log( err.message );
     }
   } );
   console.log( "Created main.cpp file" );
 
-  fs.writeFile( "./src/CMakeLists.txt", getTemplateCMakeLists( options.project, header_files, source_files ), err => {
-    if ( err ) {
-      console.log( err.message );
-    }
-  } );
+  fs.writeFile(
+    "./src/CMakeLists.txt",
+    getTemplateCMakeLists( options.project, header_files, source_files, options.config ), err => {
+      if ( err ) {
+        console.log( err.message );
+      }
+    } );
   console.log( "Created CMakefile file" );
 
-  fs.writeFile( "./src/test/test.main.cpp", getTemplateTestMain(), err => {
+  fs.writeFile( "./src/test/test.main.cpp", getTemplateTestMain( options ), err => {
     if ( err ) {
       console.log( err.message );
     }
   } );
   console.log( "Created test.main.cpp file" );
 
-  fs.writeFile( "./src/test/CMakeLists.txt", getTemplateCMakeListsTest( options.project, header_files, source_files ), err => {
-    if ( err ) {
-      console.log( err.message );
-    }
-  } );
+  fs.writeFile(
+    "./src/test/CMakeLists.txt",
+    getTemplateCMakeListsTest( options.project, header_files, source_files, options.config ), err => {
+      if ( err ) {
+        console.log( err.message );
+      }
+    } );
   console.log( "Created Unit Test CMakefile file" );
 
   utils.downloadFileHttps(