+++ changelog.md
@@ -4,6 +4,7 @@ All changes to this project will be recorded in this document.
 
 ## 0.0.8
 
+* DM-Tools now has configuration file support.
 * JavaScript template code cleanup.
 * Readme cleanup.